created a module for reading spreadsheet
