+++ files/csv.py
@@ -1,0 +1,8 @@
+from .text import ReadTextFile
+
+
+class ReadCsv(ReadTextFile):
+    pass
+    # def __init__(self, filename, mode):
+    #     self.filename = filename
+    #     self.mode = mode
