+++ usage.py
@@ -1,5 +1,11 @@
-from files.text import ReadTextFile
+# from files.text import ReadTextFile
+# from files.csv import ReadCsv
+from spreadsheet.csv_spreadsheet import OpenSpreadSheet
 
 
-def rb(files):
-    files.read_first_two()
+def rb(file):
+   return file.read_all()
+
+
+text_file = OpenSpreadSheet('iris.csv')
+print(rb(text_file))
